refactor: Update airport lookup module name
